Move helpers into src/utils
