+++ configureWorkspace/configure.ts
@@ -12,8 +12,8 @@ import { IActionContext, TelemetryProperties } from 'vscode-azureextensionui';
 import * as xml2js from 'xml2js';
 import { quickPickWorkspaceFolder } from '../commands/utils/quickPickWorkspaceFolder';
 import { ext } from '../extensionVariables';
-import { globAsync } from '../helpers/async';
-import { extractRegExGroups } from '../helpers/extractRegExGroups';
+import { globAsync } from '../src/utils/async';
+import { extractRegExGroups } from '../src/utils/extractRegExGroups';
 import { Platform, PlatformOS } from '../src/utils/platform';
 import { promptForPort, quickPickOS, quickPickPlatform } from './config-utils';
 import { configureCpp } from './configure_cpp';
