+++ configureWorkspace/configure_dotnetcore.ts
@@ -6,8 +6,8 @@
 import * as assert from 'assert';
 import * as path from 'path';
 import * as semver from 'semver';
-import { extractRegExGroups } from '../helpers/extractRegExGroups';
-import { isWindows, isWindows10RS3OrNewer, isWindows10RS4OrNewer, isWindows10RS5OrNewer } from '../helpers/osVersion';
+import { extractRegExGroups } from '../src/utils/extractRegExGroups';
+import { isWindows, isWindows10RS3OrNewer, isWindows10RS4OrNewer, isWindows10RS5OrNewer } from '../src/utils/osVersion';
 import { Platform, PlatformOS } from '../src/utils/platform';
 import { getExposeStatements, IPlatformGeneratorInfo, PackageInfo } from './configure';
