+++ extension.bundle.ts
@@ -25,7 +25,7 @@ export { IKeytar } from './src/utils/keytar';
 export { throwDockerConnectionError, internal } from './explorer/utils/dockerConnectionError';
 export { getImageLabel } from './explorer/models/getImageLabel';
 export { trimWithElipsis } from './explorer/utils/utils';
-export { isWindows10RS3OrNewer, isWindows10RS4OrNewer, isWindows10RS5OrNewer } from "./helpers/osVersion";
+export { isWindows10RS3OrNewer, isWindows10RS4OrNewer, isWindows10RS5OrNewer } from "./src/utils/osVersion";
 export { LineSplitter } from './debugging/coreclr/lineSplitter';
 export { CommandLineBuilder } from './debugging/coreclr/commandLineBuilder';
 export { DockerClient } from './debugging/coreclr/dockerClient';
@@ -39,7 +39,7 @@ export { PlatformOS, Platform } from './src/utils/platform';
 export { DockerBuildImageOptions } from "./debugging/coreclr/dockerClient";
 export { compareBuildImageOptions } from "./debugging/coreclr/dockerManager";
 export { configure, ConfigureApiOptions, ConfigureTelemetryProperties } from './configureWorkspace/configure';
-export { globAsync } from './helpers/async';
+export { globAsync } from './src/utils/async';
 export { httpsRequestBinary } from './src/utils/httpRequest';
 export { DefaultTerminalProvider } from './commands/utils/TerminalProvider';
 export { docker } from './commands/utils/docker-endpoint';
